requirements.txt
# conda create -n allostery python=3.11
--extra-index-url https://download.pytorch.org/whl/cu118
numpy == 1.26.4
torch == 2.2.2+cu118
torchvision == 0.17.2+cu118
torchaudio == 2.2.2+cu118
transformers[torch] == 4.41.2
accelerate == 0.31.0
pytorch-lightning == 2.2.5
pytorch-nlp == 0.5.0
torchmetrics == 1.4.0.post0
ipykernel == 6.29.4
jupyterlab == 4.2.2
pandas == 2.2.2
scikit-learn == 1.5.0
seqeval == 1.2.2
xgboost == 2.1.0
autogluon == 1.1.1
biopython == 1.84
bertviz == 1.4.0
ipywidgets == 8.1.3
